requirements.cloudrun.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
rembg==2.0.50
pillow==10.1.0
onnxruntime==1.16.3
numpy<2
